add UnVLV
Small EXE modification utility which replaces the MS-DOS stub with the
standard MS LINK one.

The intention is to use this on unpacked steam executables in order to
get something closer to the original. The rich header[1] is not restored
(for obvious reasons) and will be removed, if you use this on the
original exe.

[1]: http://bytepointer.com/articles/the_microsoft_rich_header.htm